Backend/fastapi_service/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
requests # For fetching data from the Node.js API
python-dotenv   